Feat: Finish The Planner and Start The Glutton
Made a decent push of progress on the Glutton path and completely polished off the Planner path.

The upgrades of ESPortions, More Mice were reconnected from their old "increase_speed" and "increase_max_fruits" names. These changes make adding many upgrades start to feel like less of a chore and more of a genuine system that's been built.

There are still a couple upgrades and the last one for Glutton won't be useful for a while, but at the least, will push the difficulty parameters in the right direction.

Fixed a ton of bugs with the fruit spawning, but got the golden fruit upgrade working properly. All in all, good session all damn day, but man this shit takes forever!