Feat: emergency push
Got one really clean run off with the Juke ability.

Took forever to get moving, but it perfectly highlights the acrobat tree in my opinion.

You can still get a timer and buy it that way, but this provides a more: make lemonade out of lemons approach. idk but it was cool so PUSH