Emergency Commit
Need to get a push in before I creash.

Made a few changes to the upgrade menu functionality. Player now presses u when they have at least 1 SP and they can access the upgrade menu.

Also added a cover up animation to cover up the screen before the upgrade menu shows up

Need to change the text that pops up, don't need it, but wanted to test it

That is all, started doing the architect path and got sidetracked trying to make a spiral animation. Will come back to later