Refactor: UI buttons and transitions
Massive overhaul on the UI system. Found a better random system for transitions to make it feel less one note.

Update functions in SceneTransition to be able to set which transition or random you want as an argument. The drip and diagonal are default randoms with a nice fast spiral for our game over and start game screen wipes

Messed with buttons for a while on the menus and got them to be very blocky, very nice

With these updates, the whole system, along with the ugprade menu pause and transition have the game feeling much more polished and tuned.