UI touch ups, that is all
Made some touch up to the UI that needed to get saved just in case. Might be tapped for the day. Currently wiped and only got a couple upgrades in.

Will be back tomorrow or hopefully later tonight with the 3 keystones and pocked garden in the next push

Fixed resume button to be a faded yellow block, it looks fine, double check wiring, I did not