Feat: Implement Shatter Reality
Implemented the entire upgrade and realized that it isn't as nice as just seemlessly swapping between quadrants.

Refactored a lot of things several times over, but figured out how to have the game boundary and camera scale with the upgrades.

Original idea for the portals seemed cool at first, but the seemless transition seems to be better regardless. That and the collision engine was messing up somewhere that I do not know.

Overall, long, long day but got this one keystone ugprade figured out, and if we even make it up to 5x5, i.e. (25 40x30) play areas, that could be fun for endless runs, idk just saying. it would ba a lot of space to not have any fruit so i'd have to figure that out, but until then, push...