Started Wiring Up Illusionist Path
Started hooking everything up, I gotta take a break but move a lot around and wanted to get it saved.

Moved Phase Shift to Illusionist and took Ghost tail from Planner and made it the base upgrade of Illusionist. The path will be the following.

Ghost Tail and Phase shift can be bought anytime but ghost tail is the base.

Blink will be a temporary teleport getaway where you'll be invincible for 3 moves to any obstacle. Then 3-card monty will be a passive upgrade that will reduce all upgrades in the shop by 1 SP, as part of the 'illusion' as it were.

The keystones will be Fractured self and dazzle camouflage, thinking about renaming it just dazzle but idk. Dazzle is just a visual trippy visual effect kind of similar to master's blueprint if you have committed to the illusionist path. Fractured self is a fun upgrade where your body becomes fragmented and you can pass through any part of the opening. Every 3 body segments will be a blank window you can pass through as a way to truly make this ugprade path distinct and feel in style.

Will start wiring in a bit. PUSH