feat: Implement Full Geomancer Path
The four main paths for the Geomancer are now setup.

The player will have a choice on how to shackle themselves with rocks with even an eventual payoff with calculated risk if you buy the Geological Survey ugprade. Very juice.

It's not wildly inventive, but being able to eat rocks feels about right, and different boosts are cool depending on the situation.

This marks the full completion of another upgrade path! Great work everyone! Only a few to go...lol as if

PUUUUUUUUUUUUUSH