Huge Refactor
Decided to add 15 difficulties and mess with the garden transition system tonight.

Added a cool mosaic, flickering fade in effect for the lvl up and a cool transition to the menu where your bonuses are displayed.

The current campaign has shifted from 5 gardens to embarking to take on 13. A new currency has been theorized. This currency is based on built up scales and will unlock things such as abilities and other things. The animations look good for now, but I got upgrade fatigue.

Now that I have animation fatigue, the next step would seem to be figure out ability slot UI, button mapping to 1-0, and then scales, and checking every ability upgrade and all that ugh. we can do it

PUSHHHHHHHHHHHH