Feat: Got the new flow working!
Huge success getting the new transitions to look at least halfway decent.

Took almost 12 hours of work but the system is looking good. We recalibrated colors so the player will be in control after a few runs. Still need to test how it works with master's blueprint and dazzle pie, but I think we're at least 50% setup for that proactively.

Fixed a small bug where the score counter was going off cuz the player would go from 0 to 1 at the start, we adjusted for this.

Made a bunch of refactors on the in between garden flow and it made me freakin' almost do it man...it was freakin'...ugh

The 4 pillars are in though. The rotating items will be a feat in it of themselves. Going to think them out more before blindly throwing them in, but there are dictionaries for them in the GameManager.

Wanted to make this push becuase it is a CLEAN loop right now and it has NOT been for the last couple days so I just wanted to get this down

PUUUUUUUUUUUUUUSH