feat: Implement upgrade nodes fully and first tab almost done with full custom icons
I'll make a bigger update once they're all done.

Still have stats and snake eyes obvi, but new nodes are very sick. Cool different color outlines as well. Need to connect the tree somehow, but maybe that's a later UI issue. After that, going to try to implement a stock ticker that pops up in the top row of the screen in the upgrade menu or on transitions. it says weird messages and other things, but just to give something during the transition. Or ever just a picture i guess, idk.

Lotta art, sorry if LFS doesn't pick it up