feat: Implement full new upgrade menu
Completed the overhaul of the upgrade menu.
The snake eyes buttons are still basic blocky for the 3 games, but the block market and Coin Flip Curious AND Passive Income all have icons

The Ssscale path is fully complete as well as the other 3.

About to do an overhaul of the pulpsicle stand to include these new nodes as well to save space and get some more ideas.

Each upgrade's stylebox border is tied to the color for the branch. It looks really nice when the color is included in the icon.

That is all...

PUUUUSHHH