feat: Implement Scrolling Headline in the Upgrade Menu
Added a hypnotic scrolling double scrolling headline into the upgrade menu experience.

Several stats are listed and I came up with about 10 headlines and 15 different things to display for stats, but it looks pretty clean as is. Just have to flesh it out and think about those people that overplay, but that can always be anytime.

Forgot if I mentioned it, but the Pulpsicle stand is fully integrated with the new icon "node" system.

The styleboxes are still a little confusing in code, but it's pretty sick when it works, but often causes more headaches when I think I found a new shortcut.

Also fixed Mulligan Munchie and Fruit spawning bugs that were being funky. Working as intended. Double check pocket garden, last time I tried it everything crashed, but I refuse to believe that's still the case lol PUUUUUUUUUUUUUSSSSSSSSSHHHHHHH