Rework: Upgrade Data Cost Overhaul and Level Progression
Made some slight tune ups to the leveling system. Juice rewarded is different as I may have mentioned before, but it operates under the x -> x+1 = +x mL juice. This way the liquid assets play is not completely broken, but instead offers at most a 50% bonus to the player.

Reworked a lot of costs with this. Not counting in gambles, gold fruit and other things pulp to juice things, the player will have about 1100 juice at the end of garden 8 and about 1600 by the end of garden 9.

I simplified and updated a lot of the abilities to go up to 9 charges at a ridiculous scale that could be afforded in a late game run when gardens are longer.

Need to add more stuff to boost juice if we go this route, also thinking of swithing juice to L instead of mL, but also idk i kinda like it. anyway, puuuuuuuuuuuuuuuuuuuuush