feat: Fully Implemented the Fang Fund sans QoL features
After adding the class grid in Class Acquisitions Tab (name pending), we now have 2 other tabs.

We have the chroma lab where you can buy colors, patterns, and backgrounds!

In addition we added a Serpent's Sigil Tab (name pending) and there are 2 categories for frames and avatars. Going to add more sections for GPS stuff and the background banner / calling card.

Fixed a weird bug with the new SaveManager where the default data was working on a corrupted file with old save stuff. The new logic is much more robust and although we need a better way to store the saves, it is working and functional right now.

Also modified the description panel to work in the fang fund. I added "name" and "description" just to avoid the headache of generating them with code. Going to work on the Shed next where you can equip this stuff and then unlock in game with chroma scales! PUUUUUUUUUUUUUUUUUUUUUUUUUUUSSSSSSSSSSHHHHHHHHH