feat: Shed Customization Tab Bugs Fixed
Fixed bugs with resizing with the upgrade nodes. They aren't perfect, but for the time being are pleasing enough on the eye to feel the loop. The shop is cool, it's easy to move around. It'd be nice for clikcing on tabs if there was some sort of effect of maybe a way to add some movement on the screen just so it doesn't look stale.

Didn't add too much, the areas with only 2 cosmetics' buttons are very long on the y-axis, this will be fixed once we add more, it's just the sizing of the container

PUSH