feat: finished fixing bugs in Shed for first few cosmetics
Fixed the update item display and added 4 default colors for the player to choose from.

Off rip, we will give the player 2 patterns to choose from. Solid Snake (just 1 color default) and simple ABAB alternating pattern with 2 colors. We will add more interesting patterns later, but this will be a good start.

About 18 colors can be bought in the shop, no gradients yet.

Avatars right now include just the old game logo and all the upgrade avatars, but it will probably include all the upgrades and some customs as well. If the player wants them.

Then there is no implementation with backgrounds. I was thinking maybe a few basic background colors to change the base game background, but grids and stuff sans masters blueprint will be tough to figure out. Perhpas it should be reserved for master's blueprint or dazzle pie specifically, unsure.

Frames are not implemented either, but are selectable!

PUUUUUUUUUUUSSHHH oh yeah and the body color 1, body color 2, etc all update after clicking on it and then clicking on the next color.