Fix: Bug Fixes and Dev Testing Gone Poorly
Fixed a ton of bugs after a few days of testing:

   Pulp not getting added from end of garden bonus  - FIXED
   Fixed Mulligan Munchie on start, now works properly: head will be purple to indicate a mistake and a new run
   Fixed score label on martyrdom
   Fractured self division no longer int on float
   Update description panel to update onp ress. Really big jitter on one frame still, but working as intended functionally.

We still have an issue with ghost tail spawning and causing a flicker on the ghost tail. This is a decent indicator for when your growth is done, but with passive gps is quite jittery. Need to come back to, spent 3 hours on this one bug today and barely got it to not look horrendous. Will come back to.

Took some time off from the programming to get some art ideas and reduce burnout. After a couple days of getting my ass kicked, it's good to be back, and for the most part, the Idle path is completely tested and working with each class properly besides the ghost tail, but even then, it's a growth issue, not a gps issue, PUUUSSSSHHH GOOD SHIT