Hot-Fix: Ghost Tail on GPS now fully fixed
Only a couple minor changes, but the ghost tail flicker is now gone.
The ghost tail still changes color upon eating a fruit, but for the time being, made a design choice to leave it as is.
That is all, also messed with a few other things but ctrl+z'd them after I went too far, just pushing good ghost build as is, good work getting it done before the wedding.

Current goals are to spend next week testing more of the bugs and incorporating at least a couple random icons to change up the flow of the day. In addition, going to try and have a public build of the game ready to be tested by labor day weekend. Ambitious, yes, but with WMU folks not coming back to town for another 3 weeks, I am going to use the time to focus on getting my daily routine in check before we start hittting night shifts for moolah.

It would be nice to be able to hire an artist with some of the dashing money and then I could save time making the art. HAHAHAHA jk, imagine PUUUSHHHHHHHHHHHHH people are so lame sometimes lol