Art Added: All pacts/difficulty art is complete
All pacts and art for them are realized and complete. Next step will be class art and figuring out the other unlockable cosmetics, but for now the selection menu is starting to look a lot more like where we want it to be.

The base pacts still have the dice so I may change that later, no code has been updated except for the references for the pngs and had to fix a couple bugs with Magician and Illusionist being the same thing and my code not realizing that. Sticking with Illusionist for the foreseeable future.

Next, trying to finish the classes. After that I will start on the milestones and other things to get a break from the art as I believe that can be finished in a day whereas all this art just takes so much love and time.

That being said, we back baby, love you Nee Nee. Wish I coulda shown you this game 7 years ago when you still remembered who I was...PUUUUUUSH :'(