Art Class Icons Completed and Small Bug Fixes
Finished class icons

squashed a nasty bug with the description panel resizing, it still wants to stick to its origin spot with the anchors, I figure with some testing we should be able to pinpoint exactly what was causing it, but it honestly looks better with a slight jitter on quick hovers than the big lopsided monstrosity that would come out of it

that is all, also added an icon path for the fang fund class icons so they will be shown as intended. In addition, the description panel seems to work best on a control as the fang fund has no issues.
the problem entities are mostly the canvas layer nodes.